third_party/blink/web_tests/external/wpt/pointerevents/pointerevent_click_during_capture_mouse-auxclick-expected.txt

This is a testharness.js-based test.
[FAIL] pointerdown/up at child1, capture at child1
  assert_equals: expected "pointerdown@child1,gotpointercapture@child1,pointerup@child1,lostpointercapture@child1,auxclick@child1" but got "pointerdown@child1,gotpointercapture@child1,pointerup@child1,auxclick@child1,lostpointercapture@child1"
[FAIL] pointerdown/up at child1, capture at child2
  assert_equals: expected "pointerdown@child1,gotpointercapture@child2,pointerup@child2,lostpointercapture@child2,auxclick@child2" but got "pointerdown@child1,gotpointercapture@child2,pointerup@child2,auxclick@child2,lostpointercapture@child2"
[FAIL] pointerdown at child1, pointerup at child2, capture at child1
  assert_equals: expected "pointerdown@child1,gotpointercapture@child1,pointerup@child1,lostpointercapture@child1,auxclick@child1" but got "pointerdown@child1,gotpointercapture@child1,pointerup@child1,auxclick@child1,lostpointercapture@child1"
[FAIL] pointerdown at child1, pointerup at child2, capture at child2
  assert_equals: expected "pointerdown@child1,gotpointercapture@child2,pointerup@child2,lostpointercapture@child2,auxclick@child2" but got "pointerdown@child1,gotpointercapture@child2,pointerup@child2,auxclick@child2,lostpointercapture@child2"
Harness: the test ran to completion.
